moved localize to mbe and tightened the possible methods
Use Literal to clearly specify which methods are available
